refactor(combobox): removing unnecessary visible tasks, example refactor
