+++ apps/website/src/routes/docs/headless/(components)/combobox/examples/highlighted-index.tsx
@@ -1,0 +1,60 @@
+import {
+  Combobox,
+  ComboboxControl,
+  ComboboxIcon,
+  ComboboxInput,
+  ComboboxListbox,
+  ComboboxOption,
+  ComboboxPortal,
+  ComboboxTrigger,
+  ResolvedOption,
+} from '@qwik-ui/headless';
+
+import { component$, useSignal } from '@builder.io/qwik';
+
+export default component$(() => {
+  const highlightedIndexSig = useSignal(2);
+
+  const highlightedExample = [
+    'not highlighted',
+    'not highlighted',
+    'highlighted by default!',
+    'not highlighted',
+  ];
+
+  return (
+    <>
+      <p class="text-center text-white">Third option highlighted! 🚨</p>
+      <Combobox
+        class="w-fit"
+        options={highlightedExample}
+        bind:highlightedIndexSig={highlightedIndexSig}
+      >
+        <ComboboxControl class="relative flex items-center rounded-sm border-[1px] border-slate-400 bg-[#1f2532]">
+          <ComboboxInput class="px-d2 w-fit bg-slate-900 px-2 pr-6 text-white placeholder:text-slate-500" />
+          <ComboboxTrigger class="group absolute right-0 h-6 w-6">
+            <ComboboxIcon class="stroke-white transition-transform duration-[450ms] group-aria-expanded:-rotate-180" />
+          </ComboboxTrigger>
+        </ComboboxControl>
+        <ComboboxPortal>
+          <ComboboxListbox
+            gutter={8}
+            size={true}
+            class="w-fit rounded-sm border-[1px] border-slate-400 bg-slate-900 px-4 py-2"
+            hide="escaped"
+            optionRenderer$={(option: ResolvedOption, index: number) => (
+              <ComboboxOption
+                key={option.key}
+                class="group rounded-sm border-2 border-transparent px-2 text-white hover:bg-slate-500  aria-disabled:text-slate-600 aria-disabled:hover:bg-slate-700 aria-selected:border-slate-200 aria-selected:bg-slate-500"
+                index={index}
+                resolved={option}
+              >
+                {option.label}
+              </ComboboxOption>
+            )}
+          />
+        </ComboboxPortal>
+      </Combobox>
+    </>
+  );
+});
