+++ apps/website/src/routes/docs/headless/(components)/combobox/examples/object.tsx
@@ -1,0 +1,56 @@
+import {
+  Combobox,
+  ComboboxControl,
+  ComboboxIcon,
+  ComboboxInput,
+  ComboboxLabel,
+  ComboboxListbox,
+  ComboboxOption,
+  ComboboxPortal,
+  ComboboxTrigger,
+  ResolvedOption,
+} from '@qwik-ui/headless';
+
+import { component$ } from '@builder.io/qwik';
+
+export default component$(() => {
+  type Jedi = {
+    value: string;
+    label: string;
+  };
+
+  const objectExample: Array<Jedi> = [
+    { value: 'anakin', label: 'Anakin Skywalker' },
+    { value: 'obi-wan', label: 'Obi-Wan Kenobi' },
+    { value: 'mace', label: 'Mace Windu' },
+    { value: 'yoda', label: 'Yoda' },
+  ];
+
+  return (
+    <Combobox class="w-fit" options={objectExample}>
+      <ComboboxLabel class=" font-semibold text-white">Jedi ⚔️</ComboboxLabel>
+      <ComboboxControl class="relative flex items-center rounded-sm border-[1px] border-slate-400 bg-[#1f2532]">
+        <ComboboxInput class="px-d2 w-44 bg-slate-900 px-2 pr-6 text-white placeholder:text-slate-500" />
+        <ComboboxTrigger class="group absolute right-0 h-6 w-6">
+          <ComboboxIcon class="stroke-white transition-transform duration-[450ms] group-aria-expanded:-rotate-180" />
+        </ComboboxTrigger>
+      </ComboboxControl>
+      <ComboboxPortal>
+        <ComboboxListbox
+          gutter={8}
+          class="w-44 rounded-sm border-[1px] border-slate-400 bg-slate-900 px-4 py-2"
+          optionRenderer$={(option: ResolvedOption, index: number) => (
+            <ComboboxOption
+              key={option.key}
+              class="group rounded-sm border-2 border-transparent px-2 text-white hover:bg-slate-500  aria-disabled:text-slate-600 aria-disabled:hover:bg-slate-700 aria-selected:border-slate-200 aria-selected:bg-slate-500"
+              index={index}
+              resolved={option}
+            >
+              {option.label}
+            </ComboboxOption>
+          )}
+        />
+      </ComboboxPortal>
+    </Combobox>
+  );
+});
