+++ packages/kit-headless/src/components/combobox/combobox-trigger.tsx
@@ -2,6 +2,7 @@ import {
   Slot,
   component$,
   useContext,
+  $,
   type QwikIntrinsicElements,
 } from '@builder.io/qwik';
 import ComboboxContextId from './combobox-context-id';
@@ -17,9 +18,12 @@ export const ComboboxTrigger = component$((props: ComboboxTriggerProps) => {
       <button
         {...props}
         ref={context.triggerRef}
-        onMouseDown$={() => {
-          context.isListboxOpenSig.value = !context.isListboxOpenSig.value;
-        }}
+        onMouseDown$={[
+          $(() => {
+            context.isListboxOpenSig.value = !context.isListboxOpenSig.value;
+          }),
+          props.onMouseDown$,
+        ]}
         tabIndex={-1}
         aria-haspopup="listbox"
         aria-controls={listboxId}
